I redid the work from earlier this week, containing as much work as possible in the file RegressionAndResiduals.m. It requires the data in the files mfi.mat and kd.mat to run. mfi.mat contains the matrix of data originally given to us by Dr. Lux, and kd.mat contains the Kd values we found hunting through the literature. For a reason explained in RegressionAndResiduals.m, I had to write the function "Error" in its own document, Error.m, as opposed to writing it as an anonymous function within RegressionAndResiduals.m.
RegressionAndResiduals.m saves the values of receptor expression found using the regression in the file R.mat. It also creates a 3-D bar graph of the differences in the MFIs that "should have" resulted from these expression levels and normalized versions of the MFI's Dr. Lux gave us, and it also creates a .csv file of these differences.